Refactor: move copy-offload logic to copyoffload_migration.py
Address myakove review comments by using callback pattern to keep
general migration code free of copy-offload specific logic.

Changes:
1. Move PVC_NAME_LABEL and capture_populate_pod_logs() to
   copyoffload_migration.py where copy-offload logic belongs

2. Add create_log_capture_callback() to provide callback compatible
   with wait_for_migration_complate's on_status_poll parameter

3. Remove copy-offload specific parameters from
   wait_for_migration_complate() - now only takes plan and callback

4. Update execute_migration() to remove copy-offload specific logic
   - General function should not have feature-specific branching

5. Update copy-offload tests to use callback pattern:
   - Create callback with create_log_capture_callback()
   - Pass to wait_for_migration_complate via on_status_poll
   - Ensures logs captured before MTV cleanup deletes pods

6. Fix execute_migration_monitoring_populator_inflight() to capture
   logs for test_check_xcopy_used verification step

7. Update AGENTS.md to document callback pattern for copy-offload
   tests instead of execute_migration() calling plan secret wait

This keeps general migration code (mtv_migration.py) free of
copy-offload branching per "Feature-Specific Logic Must Not Pollute
General Code (MUST)" rule.
